Add primitives module for Substrate test helpers
