+++ primitives/substrate/src/lib.rs
@@ -17,3 +17,6 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 pub mod justification;
+
+#[cfg(test)]
+pub mod test_helpers;
